fix: add address to receive screen, android style fixes
