+++ src/components/profile/Profile.js
@@ -8,14 +8,12 @@ import { Section, Text, Wrapper } from '../common'
 
 import UserAvatar from '../common/view/UserAvatar'
 import { withStyles } from '../../lib/styles'
-import { getDesignRelativeHeight, getDesignRelativeWidth } from '../../lib/utils/sizes'
+import { getDesignRelativeWidth } from '../../lib/utils/sizes'
 import RoundIconButton from '../common/buttons/RoundIconButton'
-import useProfile, { usePublicProfile } from '../../lib/userStorage/useProfile'
+import { usePublicProfile } from '../../lib/userStorage/useProfile'
 import { theme } from '../theme/styles'
-import Avatar from '../common/view/Avatar'
 import { useUserStorage, useWallet } from '../../lib/wallet/GoodWalletProvider'
-import { truncateMiddle } from '../../lib/utils/string'
-import { useClipboardCopy } from '../../lib/hooks/useClipboard'
+import AddressRow from '../common/view/AddressRow'
 
 import EditProfile from './EditProfile'
 import ProfileDataTable from './ProfileDataTable'
@@ -24,64 +22,6 @@ import VerifyEdit from './VerifyEdit'
 import VerifyEditCode from './VerifyEditCode'
 
 const avatarSize = getDesignRelativeWidth(136)
-
-const copiedActionTimeout = 2000 // time during which the copy success message is displayed
-
-const ProfileAvatar = withStyles(() => ({
-  avatar: {
-    borderWidth: 0,
-    backgroundColor: 'transparent',
-  },
-}))(({ styles, style }) => {
-  const { smallAvatar: avatar } = useProfile()
-
-  return <Avatar source={avatar} style={[styles.avatar, style]} imageStyle={style} unknownStyle={style} plain />
-})
-
-const AddressRow = ({ title, address = undefined, styles }) => {
-  const [performed, setPerformed] = useState(false)
-
-  const _onCopied = useCallback(() => {
-    setPerformed(true)
-    setTimeout(() => setPerformed(false), copiedActionTimeout)
-  }, [setPerformed])
-
-  const copyToClipboard = useClipboardCopy(address, _onCopied)
-
-  const handleCopy = useCallback(() => {
-    copyToClipboard()
-  }, [copyToClipboard])
-
-  const headerCopy = title === 'Wallet' ? t`My Wallet Address` : t`My Face Record ID`
-
-  const truncAddress = truncateMiddle(address, 29)
-
-  return (
-    <View style={styles.addressRow}>
-      {title === 'Wallet' ? (
-        <RoundIconButton
-          onPress={handleCopy}
-          iconSize={24}
-          iconName="wallet_alt-1"
-          style={[styles.iconContainer, { backgroundColor: theme.colors.lessDarkGray }]}
-        />
-      ) : (
-        <ProfileAvatar style={{ width: 42, height: 42 }} />
-      )}
-
-      <View style={{ flexDirection: 'column', width: 232 }}>
-        <Text style={{ textAlign: 'left', fontSize: 16 }}>{headerCopy}</Text>
-        <Text style={{ textAlign: 'left', fontSize: 14 }}>{truncAddress}</Text>
-      </View>
-      <RoundIconButton
-        onPress={handleCopy}
-        iconSize={performed ? 16 : 24}
-        iconName={performed ? 'success' : 'copy'}
-        style={styles.iconContainer}
-      />
-    </View>
-  )
-}
 
 const ProfileWrapper = ({ screenProps, styles }) => {
   const profile = usePublicProfile()
@@ -120,12 +60,12 @@ const ProfileWrapper = ({ screenProps, styles }) => {
         <View style={styles.emptySpace} />
         <ProfileDataTable profile={profile} showCustomFlag />
 
-        <Section grow justifyContent="flex-end" style={{ marginBottom: 16 }}>
-          <Section.Row>
-            <AddressRow title="Wallet" address={goodWallet.account} styles={styles} />
+        <Section grow justifyContent="flex-end" style={{ marginBottom: 16, width: '100%', padding: 0, margin: 0 }}>
+          <Section.Row style={{ width: '100%' }}>
+            <AddressRow title="Wallet" address={goodWallet.account} />
           </Section.Row>
           <Section.Row>
-            <AddressRow title="FaceId" address={faceRecordId} styles={styles} />
+            <AddressRow title="FaceId" address={faceRecordId} />
           </Section.Row>
         </Section>
       </Section>
@@ -196,31 +136,10 @@ const getStylesFromProps = ({ theme }) => {
       position: 'relative',
       zIndex: 1,
       height: avatarSize / 2,
+      margin: 0,
     },
     userName: {
       marginTop: theme.sizes.default,
-    },
-    iconContainer: {
-      height: 42,
-      width: 42,
-      backgroundColor: theme.colors.primary,
-      display: 'flex',
-      justifyContent: 'center',
-      alignItems: 'center',
-      marginBottom: getDesignRelativeHeight(4, false),
-      marginRight: 'auto',
-      marginLeft: 'auto',
-    },
-    addressRow: {
-      display: 'flex',
-      justifyContent: 'space-between',
-      flexDirection: 'row',
-      gap: 8,
-      width: '100%',
-      padding: 8,
-      backgroundColor: theme.colors.lightestGray,
-      borderRadius: 5,
-      marginBottom: 16,
     },
   }
 }
